Install system EDL files with the SDK
Install system EDL files with the SDK and provide a build flag to OESDK
to compile the SDK with/without builtin ecalls and ocalls. If this flag is
disabled, each enclave will need to explicitly import the ecalls and
ocalls used by their application in their own EDL file.

This also adds a test to show how ecalls and ocalls should be imported
when the COMPILE_SYSTEM_EDL flag is disabled.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>