Dynamically load sgx_dcap_ql instead of linking against.
This PR includes just the minimal set of changes required to
dynamically load sgx_dcap_ql instead of linking it.
Doing so allows OE SDK to be installed (as opposed to building from source)
on a non SGX machine and then run in simulation mode.
It also allows OE SDK to be installed on a non FLC machine and
then run in hardware mode (minus attestation).

This PR allows getting rid of LINK_SGX_DCAP_QL preprocessor define that
is littered around int code as well as the OEHOSTMR preprocessor define.
Both these defines are mainly aimed at avoiding the link dependency
on sgx_dcap_ql.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>