Security test - Pass enclave ptr back in to enclave
OE SDK marshalling strategy is different from Intel SGX SDK.
The host is expected to searialize all inputs before sending it
over to the enclave. This is so that the generated code works
without modification on OP-TEE where the enclave cannot access
host memory.

In Intel SDK:
   int* ptr = security_get_secret_ptr(enclave);
   security_ecall_test1(ptr); <-- will return SGX_INVALID_PARAMETER

In OE SDK,
   int* ptr = security_get_secret_ptr(enclave);
   security_ecall_test1(ptr); <-- host marshalls ptr and will obtain
                                  junk value which is sent to the enclave

all the marshalling is expected to be done by the host. Host will
read ptr which will be protected by SGX hardware.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>